TEST-33 Abstracted out common logic to retrieve a file from annotation information representing a source into SourceType which was relocated in the dominion project.
